+++ src/services/producer.rs
@@ -1,8 +1,9 @@
 use std::{str::FromStr, sync::Arc, time::Duration};
 
 use eyre::Context as _;
+use tokio::select;
 
-use crate::{consumer::QueueKey, context::Context, data};
+use crate::{context::Context, data, services::consumer::QueueKey};
 
 const INITIAL_HISTORY_POINT: u64 = 1;
 
@@ -83,10 +84,15 @@ async fn produce_events(ctx: Arc<Context>, start_from: u64, actual: u64) -> eyre
 
         tracing::debug!(mode, history_point, "Getting events...",);
 
-        let events = ctx
-            .icp()
-            .get_events(history_point)
-            .await
+        let events = select! {
+            res = ctx.icp().get_events(history_point) => res,
+            _ = ctx.cancelled() => {
+                tracing::info!(mode, history_point, "Received cancel signal, stopping...");
+                return Ok(());
+            }
+        };
+
+        let events = events
             .wrap_err_with(|| format!("Failed to get event on history_point: {history_point}"))?;
 
         tracing::debug!(mode, history_point, "Got {} events", events.len());